remove hard coded literals and namespace
This commit includes the following changes:
- new header file for physical constants
- new structure for global model parameters
- new input file to allow for runtime definition of model parameters
- new initialization functions for options, global parameters, and model parameters
- removal of most, but not all, hard coded literals.
- new physics.c file in vic_run/src to handle frequently used physical computations
- new shared print library
- new global parameter options for blowing snow
- removal of model parameters from global parameter file

Uncrustify was run at the end of the changes.

The following commit will address a memory error in the model.

+++ drivers/shared/src/calc_root_fraction.c
@@ -66,7 +66,7 @@ calc_root_fractions(veg_con_struct  *veg_con,
         zone = 0;
 
         while (zone < options.ROOT_ZONES) {
-            Zstep = (double)veg_con[veg].zone_depth[zone];
+            Zstep = veg_con[veg].zone_depth[zone];
             if ((Zsum + Zstep) <= Lsum && Zsum >= Lsum - Lstep) {
                 /** CASE 1: Root Zone Completely in Soil Layer **/
                 sum_fract += veg_con[veg].zone_fract[zone];
